# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Research Navigator API - Comprehensive scientific intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals/{runId}:
    get:
      operationId: getProposals
      tags: [proposals]
      summary: Get research proposals for a run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Research proposals
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposals:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchProposal"
                required:
                  - proposals
  /proposals/{runId}/generate:
    post:
      operationId: generateProposals
      tags: [proposals]
      summary: Generate research proposals from identified gaps
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proposals generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposals:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchProposal"
                  message:
                    type: string
                required:
                  - proposals
                  - message
...
  /report/{runId}:
    get:
      operationId: getReport
      tags: [report]
      summary: Get comprehensive research report
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comprehensive report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
  /report/{runId}/generate:
    post:
      operationId: generateReport
      tags: [report]
      summary: Generate comprehensive research report
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchReport"
components:
  schemas:
...
    ResearchProposal:
      type: object
      properties:
        id:
          type: string
        gapId:
          type: string
        title:
          type: string
        motivation:
          type: string
        researchQuestions:
          type: array
          items:
            type: string
        methodology:
          type: string
        expectedContributions:
          type: array
          items:
            type: string
        challenges:
          type: array
          items:
            type: string
        noveltyScore:
          type: number
        noveltyExplanation:
          type: string
          nullable: true
        supportingPaperIds:
          type: array
          items:
            type: string
        rawText:
          type: string
          nullable: true
      required:
        - id
        - gapId
        - title
        - motivation
        - researchQuestions
        - methodology
        - expectedContributions
        - challenges
        - noveltyScore
        - supportingPaperIds
...
    ResearchReport:
      type: object
      properties:
        runId:
          type: string
        topic:
          type: string
        overview:
          type: string
        graphInsights:
          type: string
        trendsSummary:
          type: string
        gapsSummary:
          type: string
        controversySummary:
          type: string
        recommendations:
          type: array
          items:
            type: string
        totalPapers:
          type: integer
        generatedAt:
          type: string
          format: date-time
      required:
        - runId
        - topic
        - overview
        - graphInsights
        - trendsSummary
        - gapsSummary
        - controversySummary
        - recommendations
        - totalPapers
        - generatedAt